split parser.h into parser.cpp/h
