+++ runtime/msgpack/parser.cpp
@@ -7,11 +7,12 @@
 //    (See accompanying file LICENSE_1_0.txt or copy at
 //    http://www.boost.org/LICENSE_1_0.txt)
 //
-#pragma once
 
 #include <cstddef>
 #include <vector>
 
+#include "runtime/msgpack/object_visitor.h"
+#include "runtime/msgpack/parser.h"
 #include "runtime/msgpack/parse_return.h"
 #include "runtime/msgpack/unpack_decl.h"
 #include "runtime/msgpack/unpack_exception.h"
@@ -55,55 +56,6 @@ template<typename T>
 inline typename std::enable_if<sizeof(T) == 8>::type load(T &dst, const char *n) {
   _msgpack_load64(T, n, &dst);
 }
-
-enum {
-  MSGPACK_CS_HEADER = 0x00, // nil
-
-  // MSGPACK_CS_                = 0x01,
-  // MSGPACK_CS_                = 0x02,  // false
-  // MSGPACK_CS_                = 0x03,  // true
-
-  MSGPACK_CS_BIN_8 = 0x04,
-  MSGPACK_CS_BIN_16 = 0x05,
-  MSGPACK_CS_BIN_32 = 0x06,
-
-  MSGPACK_CS_EXT_8 = 0x07,
-  MSGPACK_CS_EXT_16 = 0x08,
-  MSGPACK_CS_EXT_32 = 0x09,
-
-  MSGPACK_CS_FLOAT = 0x0a,
-  MSGPACK_CS_DOUBLE = 0x0b,
-  MSGPACK_CS_UINT_8 = 0x0c,
-  MSGPACK_CS_UINT_16 = 0x0d,
-  MSGPACK_CS_UINT_32 = 0x0e,
-  MSGPACK_CS_UINT_64 = 0x0f,
-  MSGPACK_CS_INT_8 = 0x10,
-  MSGPACK_CS_INT_16 = 0x11,
-  MSGPACK_CS_INT_32 = 0x12,
-  MSGPACK_CS_INT_64 = 0x13,
-
-  MSGPACK_CS_FIXEXT_1 = 0x14,
-  MSGPACK_CS_FIXEXT_2 = 0x15,
-  MSGPACK_CS_FIXEXT_4 = 0x16,
-  MSGPACK_CS_FIXEXT_8 = 0x17,
-  MSGPACK_CS_FIXEXT_16 = 0x18,
-
-  MSGPACK_CS_STR_8 = 0x19,  // str8
-  MSGPACK_CS_STR_16 = 0x1a, // str16
-  MSGPACK_CS_STR_32 = 0x1b, // str32
-  MSGPACK_CS_ARRAY_16 = 0x1c,
-  MSGPACK_CS_ARRAY_32 = 0x1d,
-  MSGPACK_CS_MAP_16 = 0x1e,
-  MSGPACK_CS_MAP_32 = 0x1f,
-
-  // MSGPACK_ACS_BIG_INT_VALUE,
-  // MSGPACK_ACS_BIG_FLOAT_VALUE,
-  MSGPACK_ACS_STR_VALUE,
-  MSGPACK_ACS_BIN_VALUE,
-  MSGPACK_ACS_EXT_VALUE
-};
-
-enum msgpack_container_type { MSGPACK_CT_ARRAY_ITEM, MSGPACK_CT_MAP_KEY, MSGPACK_CT_MAP_VALUE };
 
 template<typename Visitor>
 struct array_sv {
@@ -159,154 +111,115 @@ private:
   Visitor &visitor_;
 };
 
-struct unpack_stack {
-  struct stack_elem {
-    stack_elem(msgpack_container_type type, uint32_t rest) noexcept
-      : m_type(type)
-      , m_rest(rest) {}
-    msgpack_container_type m_type;
-    uint32_t m_rest;
-  };
-
-  unpack_stack() noexcept {
-    m_stack.reserve(MSGPACK_EMBED_STACK_SIZE);
+template<typename Visitor>
+parse_return unpack_stack::push(Visitor &visitor, msgpack_container_type type, uint32_t rest) {
+  m_stack.emplace_back(type, rest);
+  switch (type) {
+    case MSGPACK_CT_ARRAY_ITEM:
+      return visitor.start_array_item() ? PARSE_CONTINUE : PARSE_STOP_VISITOR;
+    case MSGPACK_CT_MAP_KEY:
+      return visitor.start_map_key() ? PARSE_CONTINUE : PARSE_STOP_VISITOR;
+    case MSGPACK_CT_MAP_VALUE:
+      assert(0);
+      return PARSE_STOP_VISITOR;
   }
-
-  template<typename Visitor>
-  parse_return push(Visitor &visitor, msgpack_container_type type, uint32_t rest) {
-    m_stack.emplace_back(type, rest);
-    switch (type) {
-      case MSGPACK_CT_ARRAY_ITEM:
-        return visitor.start_array_item() ? PARSE_CONTINUE : PARSE_STOP_VISITOR;
-      case MSGPACK_CT_MAP_KEY:
-        return visitor.start_map_key() ? PARSE_CONTINUE : PARSE_STOP_VISITOR;
-      case MSGPACK_CT_MAP_VALUE:
-        assert(0);
-        return PARSE_STOP_VISITOR;
-    }
-    assert(0);
-    return PARSE_STOP_VISITOR;
-  }
-
-  template<typename Visitor>
-  parse_return consume(Visitor &visitor) {
-    while (!m_stack.empty()) {
-      stack_elem &e = m_stack.back();
-      switch (e.m_type) {
-        case MSGPACK_CT_ARRAY_ITEM:
-          if (!visitor.end_array_item())
-            return PARSE_STOP_VISITOR;
-          if (--e.m_rest == 0) {
-            m_stack.pop_back();
-            if (!visitor.end_array())
-              return PARSE_STOP_VISITOR;
-          } else {
-            if (!visitor.start_array_item())
-              return PARSE_STOP_VISITOR;
-            return PARSE_CONTINUE;
-          }
-          break;
-        case MSGPACK_CT_MAP_KEY:
-          if (!visitor.end_map_key())
-            return PARSE_STOP_VISITOR;
-          if (!visitor.start_map_value())
-            return PARSE_STOP_VISITOR;
-          e.m_type = MSGPACK_CT_MAP_VALUE;
-          return PARSE_CONTINUE;
-        case MSGPACK_CT_MAP_VALUE:
-          if (!visitor.end_map_value())
-            return PARSE_STOP_VISITOR;
-          if (--e.m_rest == 0) {
-            m_stack.pop_back();
-            if (!visitor.end_map())
-              return PARSE_STOP_VISITOR;
-          } else {
-            e.m_type = MSGPACK_CT_MAP_KEY;
-            if (!visitor.start_map_key())
-              return PARSE_STOP_VISITOR;
-            return PARSE_CONTINUE;
-          }
-          break;
-      }
-    }
-    return PARSE_SUCCESS;
-  }
-
-private:
-  std::vector<stack_elem> m_stack;
-};
+  assert(0);
+  return PARSE_STOP_VISITOR;
+}
 
 template<typename Visitor>
-class parser {
-public:
-  explicit parser(Visitor &visitor) noexcept
-    : visitor_(visitor) {}
-
-  parse_return execute(const char *data, std::size_t len, std::size_t &off);
-
-private:
-  template<typename T>
-  static uint32_t next_cs(T p) {
-    return static_cast<uint32_t>(*p) & 0x1f;
-  }
-
-  template<typename T, typename StartVisitor, typename EndVisitor>
-  parse_return start_aggregate(const StartVisitor &sv, const EndVisitor &ev, const char *load_pos, std::size_t &off) {
-    typename value<T>::type size;
-    load<T>(size, load_pos);
-    ++m_current;
-    if (size == 0) {
-      if (!sv(size)) {
-        off = m_current - m_start;
-        return PARSE_STOP_VISITOR;
-      }
-      if (!ev()) {
-        off = m_current - m_start;
-        return PARSE_STOP_VISITOR;
-      }
-      parse_return ret = m_stack.consume(visitor_);
-      if (ret != PARSE_CONTINUE) {
-        off = m_current - m_start;
-        return ret;
-      }
-    } else {
-      if (!sv(size)) {
-        off = m_current - m_start;
-        return PARSE_STOP_VISITOR;
-      }
-      parse_return ret = m_stack.push(visitor_, sv.type(), static_cast<uint32_t>(size));
-      if (ret != PARSE_CONTINUE) {
-        off = m_current - m_start;
-        return ret;
-      }
+parse_return unpack_stack::consume(Visitor &visitor) {
+  while (!m_stack.empty()) {
+    stack_elem &e = m_stack.back();
+    switch (e.m_type) {
+      case MSGPACK_CT_ARRAY_ITEM:
+        if (!visitor.end_array_item())
+          return PARSE_STOP_VISITOR;
+        if (--e.m_rest == 0) {
+          m_stack.pop_back();
+          if (!visitor.end_array())
+            return PARSE_STOP_VISITOR;
+        } else {
+          if (!visitor.start_array_item())
+            return PARSE_STOP_VISITOR;
+          return PARSE_CONTINUE;
+        }
+        break;
+      case MSGPACK_CT_MAP_KEY:
+        if (!visitor.end_map_key())
+          return PARSE_STOP_VISITOR;
+        if (!visitor.start_map_value())
+          return PARSE_STOP_VISITOR;
+        e.m_type = MSGPACK_CT_MAP_VALUE;
+        return PARSE_CONTINUE;
+      case MSGPACK_CT_MAP_VALUE:
+        if (!visitor.end_map_value())
+          return PARSE_STOP_VISITOR;
+        if (--e.m_rest == 0) {
+          m_stack.pop_back();
+          if (!visitor.end_map())
+            return PARSE_STOP_VISITOR;
+        } else {
+          e.m_type = MSGPACK_CT_MAP_KEY;
+          if (!visitor.start_map_key())
+            return PARSE_STOP_VISITOR;
+          return PARSE_CONTINUE;
+        }
+        break;
     }
-    m_cs = MSGPACK_CS_HEADER;
-    return PARSE_CONTINUE;
   }
+  return PARSE_SUCCESS;
+}
 
-  parse_return after_visit_proc(bool visit_result, std::size_t &off) {
-    ++m_current;
-    if (!visit_result) {
+template<typename Visitor>
+template<typename T, typename StartVisitor, typename EndVisitor>
+parse_return parser<Visitor>::start_aggregate(const StartVisitor &sv, const EndVisitor &ev, const char *load_pos, std::size_t &off) {
+  typename value<T>::type size;
+  load<T>(size, load_pos);
+  ++m_current;
+  if (size == 0) {
+    if (!sv(size)) {
+      off = m_current - m_start;
+      return PARSE_STOP_VISITOR;
+    }
+    if (!ev()) {
       off = m_current - m_start;
       return PARSE_STOP_VISITOR;
     }
     parse_return ret = m_stack.consume(visitor_);
     if (ret != PARSE_CONTINUE) {
       off = m_current - m_start;
+      return ret;
     }
-    m_cs = MSGPACK_CS_HEADER;
-    return ret;
+  } else {
+    if (!sv(size)) {
+      off = m_current - m_start;
+      return PARSE_STOP_VISITOR;
+    }
+    parse_return ret = m_stack.push(visitor_, sv.type(), static_cast<uint32_t>(size));
+    if (ret != PARSE_CONTINUE) {
+      off = m_current - m_start;
+      return ret;
+    }
   }
+  m_cs = MSGPACK_CS_HEADER;
+  return PARSE_CONTINUE;
+}
 
-  Visitor &visitor_;
-
-  const char *m_start{nullptr};
-  const char *m_current{nullptr};
-
-  std::size_t m_trail{0};
-  uint32_t m_cs{MSGPACK_CS_HEADER};
-  unpack_stack m_stack;
-};
+template<typename Visitor>
+parse_return parser<Visitor>::after_visit_proc(bool visit_result, std::size_t &off) {
+  ++m_current;
+  if (!visit_result) {
+    off = m_current - m_start;
+    return PARSE_STOP_VISITOR;
+  }
+  parse_return ret = m_stack.consume(visitor_);
+  if (ret != PARSE_CONTINUE) {
+    off = m_current - m_start;
+  }
+  m_cs = MSGPACK_CS_HEADER;
+  return ret;
+}
 
 template<std::size_t N>
 inline void check_ext_size(std::size_t /*size*/) {}
@@ -318,7 +231,7 @@ inline void check_ext_size<4>(std::size_t size) {
 }
 
 template<typename Visitor>
-inline parse_return parser<Visitor>::execute(const char *data, std::size_t len, std::size_t &off) {
+parse_return parser<Visitor>::execute(const char *data, std::size_t len, std::size_t &off) {
   assert(len >= off);
 
   m_start = data;
@@ -740,7 +653,7 @@ inline parse_return parser<Visitor>::execute(const char *data, std::size_t len, 
 }
 
 template<typename Visitor>
-inline parse_return parse_imp(const char *data, size_t len, size_t &off, Visitor &v) {
+parse_return parse_imp(const char *data, size_t len, size_t &off, Visitor &v) {
   std::size_t noff = off;
 
   if (len <= noff) {
@@ -766,4 +679,5 @@ inline parse_return parse_imp(const char *data, size_t len, size_t &off, Visitor
   }
 }
 
+template parse_return parse_imp<object_visitor>(const char *data, size_t len, size_t &off, object_visitor &v);
 } // namespace msgpack
